backend/requirements.txt
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.0.4
Flask==2.2.2
idna==3.4
importlib-metadata==5.1.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.5
packaging==22.0
Pillow==9.3.0
pkg_resources==0.0.0
pluggy==1.0.0
pytest==7.2.0
requests==2.28.1
tomli==2.0.1
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
